(issue14) Move message_router to respective app/helpers folder
