+++ tests/app/helpers/message_router_test.py
@@ -1,20 +1,15 @@
-import os
-import sys
-import pytest
-
-sys.path.insert(
-    0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..", ".."))
-)
-
+from app.helpers.message_router import MessageRouter
 from agents.helpers.agent_manager import AgentManager
-from agents.helpers.message_router import MessageRouter
+from agents.helpers.inter_agent_comm import InterAgentComm
+import pytest
 
 
 @pytest.mark.asyncio
 async def test_message_routing():
-    # Initialize AgentManager and MessageRouter
-    agent_manager = AgentManager()
-    message_router = MessageRouter(agent_manager)
+    # Initialize InterAgentComm, AgentManager and MessageRouter
+    inter_agent_comm = InterAgentComm()
+    agent_manager = AgentManager(inter_agent_comm)
+    message_router = MessageRouter(agent_manager, inter_agent_comm)
 
     # Register an agent for testing
     await agent_manager.register_agent(
